requirements.txt
alembic==1.4.2
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
WTForms==2.3.1
